Directory rename: clear all affected metadata, prevent double-dir
When directory was renamed and  metadata cleared() was emitted, the model
was immediately refreshed, which caused access to the old directory,
which in turn caused it to be created by metadata's QSettings. I'm not
sure why would the model even see the directory when it's supposed to be
renamed, but not firing this signal solves the problem. QFileSystemModel
is monitoring the directory for changes, so it notices the rename
anyway.