Synchronize access to metadata cache
The cache is accessed by multiple threads, namely the main thread and
the thumbnail worker thread, so the access must be synchronized using a
mutex. Furthermore, data and thumbnail includes were separated, as they
should be. Loading of data from Pro/E does not clear the metadata cache,
as there is no reason to.